fix/refactor: vision constants + forgot to add back real camera data and instance getter

+++ src/main/java/com/stuypulse/robot/subsystems/vision/VisionIO.java
@@ -50,7 +50,7 @@ public interface VisionIO {
 
     public int pipeline = 0;
 
-    public double[] aprilTagIDWhitelist = Field.ALL_TAGS;
+    public int[] aprilTagIDWhitelist = Field.ALL_TAGS;
   }
 
   public default void updateInputs(VisionIOInputs inputs) {}
